requirements.txt
# Core dependencies for poaching detection system
pyserial>=3.5
numpy>=1.24.0

# Supabase integration
supabase>=2.0.0
python-dotenv>=1.0.0

# ML Model dependencies
scikit-learn>=1.3.0
joblib>=1.3.0
librosa>=0.10.0

# Location services
requests>=2.31.0
